requirements.txt
Django==4.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
python-decouple==3.5
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0

